requirements.txt
# Core libraries
selenium
pytest

# Pytest plugins
pytest-order

# Reporting tools
allure-pytest
